+++ lib/xcodeproj/constants.rb
@@ -32,6 +32,10 @@ module Xcodeproj
     #
     LAST_SWIFT_UPGRADE_CHECK = '0700'
 
+    # @return [String] The version of `.xcscheme` files supported by Xcodeproj
+    #
+    XCSCHEME_FORMAT_VERSION = '1.3'
+
     # @return [Hash] The all the known ISAs grouped by superclass.
     #
     KNOWN_ISAS = {
@@ -104,6 +108,7 @@ module Xcodeproj
       :dynamic_library   => 'com.apple.product-type.library.dynamic',
       :static_library    => 'com.apple.product-type.library.static',
       :bundle            => 'com.apple.product-type.bundle',
+      :octest_bundle     => 'com.apple.product-type.bundle',
       :unit_test_bundle  => 'com.apple.product-type.bundle.unit-test',
       :app_extension     => 'com.apple.product-type.app-extension',
       :command_line_tool => 'com.apple.product-type.tool',
@@ -116,11 +121,13 @@ module Xcodeproj
     # @return [Hash] The extensions or the various product UTIs.
     #
     PRODUCT_UTI_EXTENSIONS = {
-      :application     => 'app',
-      :framework       => 'framework',
-      :dynamic_library => 'dylib',
-      :static_library  => 'a',
-      :bundle          => 'bundle',
+      :application      => 'app',
+      :framework        => 'framework',
+      :dynamic_library  => 'dylib',
+      :static_library   => 'a',
+      :bundle           => 'bundle',
+      :octest_bundle    => 'octest',
+      :unit_test_bundle => 'xctest',
     }.freeze
 
     # @return [Hash] The common build settings grouped by platform, and build